+++ paperforge/plugin/vitest.config.ts
@@ -5,6 +5,5 @@ export default defineConfig({
     environment: 'jsdom',
     include: ['tests/**/*.test.mjs'],
     globals: true,
-    setupFiles: ['./tests/setup.mjs'],
   },
 });
